requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.2
django-crispy-forms==2.0
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
httpie==3.2.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mssql-django==1.3
multidict==6.0.4
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyodbc==4.0.39
PySocks==1.7.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.10.3
simplejson==3.19.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
style==1.1.0
swagger-spec-validator==3.0.3
typing_extensions==4.7.0
tzdata==2023.3
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.4
